backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
modal
openai
anthropic
google-generativeai
httpx
